Extract the boot/credentials flow out of main.ts into boot-flow.ts
Pure move, no behavior change: the login/boot overlay pairs, the
saved-session auto-connect (stall watchdog + backoff retry), the
membership-card hand-off, the classic login form handlers, and the
demo-mode boot all now live in boot-flow.ts behind an initBootFlow(deps)
surface, the same shape counter-terminal-flow.ts uses. main.ts keeps the
store state (libraries, games, scene) and passes setters/loaders in.

main.ts was 79 lines from its enforced budget ceiling; this frees 465 —
room the first-run opening-day flow (#41) needs.
